+++ src/toggle_tests.sh
@@ -9,9 +9,9 @@ CURRENT_DIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
 # shellcheck source=./helpers.sh
 source "$CURRENT_DIR/helpers.sh"
 
-declare delimiter=">>>END" exit_codes=(0 0 0 0) test_name
+declare delimiter=">>>END" exit_codes=(0 0 0 0) test_name should_fail
 test_toggle() {
-	local i workdir f_call_id f_input f_output f_expected
+	local i workdir f_call_id f_args f_input f_output f_expected
 
 	# Prepare inputs
 	workdir=$(mktemp -d)
@@ -19,6 +19,7 @@ test_toggle() {
 	trap "rm -rf '$workdir'" EXIT
 
 	f_call_id="$workdir/call_id"
+	f_args="$workdir/args"
 	f_input="$workdir/iutput"
 	f_output="$workdir/output"
 
@@ -28,24 +29,38 @@ test_toggle() {
 		i=$((i + 1))
 	done
 
+	# Fake argument expansions
+	local i=1 fake_expanded_args=()
+	while [[ $i -le $# ]]; do
+		fake_expanded_args+=("t_argv_$i=${!i}")
+		i=$((i + 1))
+	done
+	fake_batch_options "${fake_expanded_args[@]}" >"$f_args"
+
 	# Do call popup-toggle
-	export delimiter f_call_id f_input f_output
-	command "$CURRENT_DIR/toggle.sh" "$@" || exit 1
+	export delimiter f_call_id f_args f_input f_output
+	export TMUX=tmux-test
+	if [[ -n $should_fail ]]; then
+		! command "$CURRENT_DIR/toggle.sh" "$@" 4>&2 2>>"$f_output" || exit 1
+		f_expected="$CURRENT_DIR/toggle_tests/$test_name.stderr"
+	else
+		command "$CURRENT_DIR/toggle.sh" "$@" 4>&2 || exit 1
+		f_expected="$CURRENT_DIR/toggle_tests/$test_name.stdout"
+	fi
 
 	# Validate outputs
-	f_expected="$CURRENT_DIR/toggle_tests/$test_name.stdout"
 	if [[ $TEST_OVERWRITE = 1 ]]; then
 		mkdir -p "$(dirname "$f_expected")"
 		cp "$f_output" "$f_expected"
 	else
-		git diff --exit-code "$f_expected" "$f_output"
+		diff -u --color=auto "$f_expected" "$f_output"
 	fi
 }
 
 # Ensure out fake executable tmux is picked at first.
 export PATH="$CURRENT_DIR/toggle_tests:$PATH"
 export SHELL="/system/shell"
-export TMUX_POPUP_SERVER=
+unset TMUX TMUX_POPUP_SERVER __tmux_popup_caller __tmux_popup_name
 
 # Force subshell to ensure modifications are temporary.
 (
@@ -56,13 +71,13 @@ export TMUX_POPUP_SERVER=
 
 (
 	test_name="close_popup"
-	export t_opened_name="p_close"
+	export __tmux_popup_name="p_close"
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_close"
 ) || exit 1
 (
 	test_name="switch_popup"
-	export t_opened_name="p_switch_1"
+	export __tmux_popup_name="p_switch_1"
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_switch_2"
 ) || exit 1
@@ -70,7 +85,7 @@ export TMUX_POPUP_SERVER=
 (
 	test_name="switch_new_popup"
 	exit_codes=(0 1 0 0)
-	export t_opened_name="p_switch_1"
+	export __tmux_popup_name="p_switch_1"
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_switch_2"
 ) || exit 1
@@ -78,7 +93,7 @@ export TMUX_POPUP_SERVER=
 (
 	test_name="force_close_popup"
 	export t_toggle_mode="force-close"
-	export t_opened_name="p_force_close_1"
+	export __tmux_popup_name="p_force_close_1"
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_force_close_2"
 ) || exit 1
@@ -86,7 +101,7 @@ export TMUX_POPUP_SERVER=
 (
 	test_name="open_nested_popup"
 	export t_toggle_mode="force-open"
-	export t_opened_name="p_open_nested_1"
+	export __tmux_popup_name="p_open_nested_1"
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_open_nested_2"
 ) || exit 1
@@ -101,7 +116,7 @@ export TMUX_POPUP_SERVER=
 (
 	test_name="open_nested_with_toggle_key"
 	export t_toggle_mode="force-open"
-	export t_opened_name="p_nested_toggle_key_1"
+	export __tmux_popup_name="p_nested_toggle_key_1"
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_nested_toggle_key_2" --toggle-key="-n M-o"
 ) || exit 1
@@ -135,6 +150,7 @@ export TMUX_POPUP_SERVER=
 
 (
 	test_name="open_with_directory"
+	should_fail=1
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_open_with_directory" -d"{popup_caller_pane_path}"
 ) || exit 1
@@ -160,8 +176,9 @@ export TMUX_POPUP_SERVER=
 
 (
 	test_name="switch_with_directory"
+	should_fail=1
 	exit_codes=(0 1 0 0)
-	export t_opened_name="p_switch_with_directory_1"
+	export __tmux_popup_name="p_switch_with_directory_1"
 	begin_test "$test_name" || exit 0
 	test_toggle --name="p_switch_with_directory_2" -d"{popup_caller_pane_path}"
 ) || exit 1
@@ -169,7 +186,7 @@ export TMUX_POPUP_SERVER=
 # ID is taken as the opened name when it is specified.
 (
 	test_name="open_nested_with_id"
-	export t_opened_name="default"
+	export __tmux_popup_name="default"
 	begin_test "$test_name" || exit 0
 	test_toggle --id="p_open_nested_with_id" --toggle-mode=force-open
 ) || exit 1